+++ engine/maths/seedprimes.cpp
@@ -38,7 +38,7 @@ namespace regina {
 // way of everything else.
 
 const size_t Primes::numPrimeSeeds = 10000;
-const unsigned long Primes::primeSeedList[] = {
+const int32_t Primes::primeSeedList[] = {
       2,      3,      5,      7,     11,     13,     17,     19,     23,     29,
      31,     37,     41,     43,     47,     53,     59,     61,     67,     71,
      73,     79,     83,     89,     97,    101,    103,    107,    109,    113,
